Consolidate TransformExecutor into step-executors
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/BelfordZ/open-rpc-flow?shareId=XXXX-XXXX-XXXX-XXXX).
