+++ src/index.ts
@@ -1,6 +1,6 @@
 export { Flow, Step, JsonRpcRequest } from './types';
 export { FlowExecutor } from './flow-executor';
-export { TransformOperation } from './transform-executor';
+export { TransformOperation, TransformExecutor } from './step-executors/transform-executor';
 export { ExpressionEvaluator } from './expression-evaluator';
 export { ReferenceResolver } from './reference-resolver';
 import metaSchemaContent from '../meta-schema.json';
